+++ react/src/components/sections/Section2B.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import Sidebar from "../Sidebar";
-import Objective2b from "./Objectives/Objective2b";
+import Objective2b from "./objectives/Objective2b";
 
 class Section2b extends Component {
   constructor(props) {
