+++ react/src/components/sections/objectives/Objective2b.js
